+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 Babashka [http-client](https://github.com/babashka/http-client): HTTP client for Clojure and babashka built on java.net.http
 
-## Unreleased
+## 0.2.9
 
 - Accept `java.net.URI` as uri directly in `request`, `get`, etc.
 - [#22](https://github.com/babashka/http-client/issues/22): Support options for `:ssl-context`, similar to hato
+- [#23](https://github.com/babashka/http-client/issues/23): ease construction of `ProxySelector` via `:proxy` key
 
 ## 0.1.8